use BaseCreateAlterUserStmt
