+++ fe/fe-core/src/main/java/com/starrocks/sql/analyzer/PrivilegeStmtAnalyzer.java
@@ -3,7 +3,6 @@ package com.starrocks.sql.analyzer;
 
 import com.google.common.base.Strings;
 import com.starrocks.analysis.AlterUserStmt;
-import com.starrocks.analysis.CreateUserStmt;
 import com.starrocks.analysis.StatementBase;
 import com.starrocks.analysis.UserIdentity;
 import com.starrocks.cluster.ClusterNamespace;
@@ -20,6 +19,7 @@ import com.starrocks.mysql.privilege.Role;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.ast.AstVisitor;
+import com.starrocks.sql.ast.BaseCreateAlterUserStmt;
 import com.starrocks.sql.ast.BaseGrantRevokeImpersonateStmt;
 import com.starrocks.sql.ast.BaseGrantRevokeRoleStmt;
 import com.starrocks.sql.ast.ExecuteAsStmt;
@@ -38,21 +38,8 @@ public class PrivilegeStmtAnalyzer {
         /**
          * analyse user identity + check if user exists in UserPrivTable
          */
-        private void analyseUserAndCheckExist(UserIdentity userIdent, ConnectContext session) {
-            // validate user
-            analyseUser(userIdent, session);
-
-            // check if user exists
-            if (!session.getGlobalStateMgr().getAuth().getUserPrivTable().doesUserExist(userIdent)) {
-                throw new SemanticException("user " + userIdent + " not exist!");
-            }
-        }
-
-        /**
-         *  analyse user identity
-         */
-        private void analyseUser(UserIdentity userIdent, ConnectContext session) {
-            // validate user
+        private void analyseUserAndCheckExist(UserIdentity userIdent, ConnectContext session, boolean checkExist) {
+            // analyse user identity
             try {
                 userIdent.analyze();
             } catch (AnalysisException e) {
@@ -60,6 +47,13 @@ public class PrivilegeStmtAnalyzer {
                 // that is permitted to throw during analyzing phrase under the new framework for compatibility.
                 // Remove it after all old methods migrate to the new framework
                 throw new SemanticException(e.getMessage());
+            }
+
+            if (checkExist) {
+                // check if user exists
+                if (!session.getGlobalStateMgr().getAuth().getUserPrivTable().doesUserExist(userIdent)) {
+                    throw new SemanticException("user " + userIdent + " not exist!");
+                }
             }
         }
 
@@ -88,7 +82,7 @@ public class PrivilegeStmtAnalyzer {
          */
         @Override
         public Void visitGrantRevokeRoleStatement(BaseGrantRevokeRoleStmt stmt, ConnectContext session) {
-            analyseUserAndCheckExist(stmt.getUserIdent(), session);
+            analyseUserAndCheckExist(stmt.getUserIdent(), session, true);
             stmt.setQualifiedRole(analyseRoleName(stmt.getRole(), session));
             return null;
         }
@@ -99,8 +93,8 @@ public class PrivilegeStmtAnalyzer {
          */
         @Override
         public Void visitGrantRevokeImpersonateStatement(BaseGrantRevokeImpersonateStmt stmt, ConnectContext session) {
-            analyseUserAndCheckExist(stmt.getAuthorizedUser(), session);
-            analyseUserAndCheckExist(stmt.getSecuredUser(), session);
+            analyseUserAndCheckExist(stmt.getAuthorizedUser(), session, true);
+            analyseUserAndCheckExist(stmt.getSecuredUser(), session, true);
             return null;
         }
 
@@ -112,52 +106,15 @@ public class PrivilegeStmtAnalyzer {
             if (stmt.isAllowRevert()) {
                 throw new SemanticException("`EXECUTE AS` must use with `WITH NO REVERT` for now!");
             }
-            analyseUserAndCheckExist(stmt.getToUser(), session);
-            return null;
-        }
-
-        @Override
-        public Void visitAlterUserStatement(AlterUserStmt stmt, ConnectContext session) {
-            analyseUserAndCheckExist(stmt.getUserIdent(), session);
-            /*
-             * IDENTIFIED BY
-             */
-            // convert password to hashed password
-            stmt.setScramblePassword(
-                    analysePassword(stmt.getUserIdent(), stmt.getOriginalPassword(), stmt.isPasswordPlain()));
-
-            /*
-             * IDENTIFIED WITH
-             */
-            if (!Strings.isNullOrEmpty(stmt.getAuthPlugin())) {
-                stmt.setAuthPlugin(stmt.getAuthPlugin().toUpperCase());
-                if (AuthPlugin.AUTHENTICATION_LDAP_SIMPLE.name().equals(stmt.getAuthPlugin())) {
-                    stmt.setUserForAuthPlugin(stmt.getAuthString());
-                } else if (AuthPlugin.MYSQL_NATIVE_PASSWORD.name().equals(stmt.getAuthPlugin())) {
-                    // in this case, authString is password
-                    stmt.setScramblePassword(analysePassword(stmt.getUserIdent(), stmt.getAuthString(),
-                            stmt.isPasswordPlain()));
-                } else if (AuthPlugin.AUTHENTICATION_KERBEROS.name().equalsIgnoreCase(stmt.getAuthPlugin()) &&
-                        GlobalStateMgr.getCurrentState().getAuth().isSupportKerberosAuth()) {
-                    // In kerberos authentication, userForAuthPlugin represents the user principal realm.
-                    // If user realm is not specified when creating user, the service principal realm will be used as
-                    // the user principal realm by default.
-                    if (stmt.getAuthString() != null) {
-                        stmt.setUserForAuthPlugin(stmt.getAuthString());
-                    } else {
-                        stmt.setUserForAuthPlugin(Config.authentication_kerberos_service_principal.split("@")[1]);
-                    }
-                } else {
-                    ErrorReport.reportSemanticException(ErrorCode.ERR_AUTH_PLUGIN_NOT_LOADED, stmt.getAuthPlugin());
-                }
-            }
+            analyseUserAndCheckExist(stmt.getToUser(), session, true);
             return null;
         }
 
         /**
          * convert password to hashed password
          */
-        private byte[] analysePassword(UserIdentity userIdent, String originalPassword, boolean isPasswordPlain) {
+        private byte[] analysePassword(UserIdentity userIdent, String originalPassword, boolean isPasswordPlain,
+                                       boolean checkReuse) {
             if (Strings.isNullOrEmpty(originalPassword)) {
                 return new byte[0];
             }
@@ -165,7 +122,9 @@ public class PrivilegeStmtAnalyzer {
                 if (isPasswordPlain) {
                     // plain password should check for validation & reuse
                     Auth.validatePassword(originalPassword);
-                    GlobalStateMgr.getCurrentState().getAuth().checkPasswordReuse(userIdent, originalPassword);
+                    if (checkReuse) {
+                        GlobalStateMgr.getCurrentState().getAuth().checkPasswordReuse(userIdent, originalPassword);
+                    }
                     // convert plain password to scramble
                     return MysqlPassword.makeScrambledPassword(originalPassword);
                 } else {
@@ -179,26 +138,26 @@ public class PrivilegeStmtAnalyzer {
             }
         }
 
-        @Override
-        public Void visitCreateUserStatement(CreateUserStmt stmt, ConnectContext session) {
-            analyseUser(stmt.getUserIdent(), session);
+        public Void visitCreateAlterUserStmt(BaseCreateAlterUserStmt stmt, ConnectContext session) {
+            analyseUserAndCheckExist(stmt.getUserIdent(), session, stmt instanceof AlterUserStmt);
             /*
              * IDENTIFIED BY
              */
             stmt.setScramblePassword(
-                    analysePassword(stmt.getUserIdent(), stmt.getOriginalPassword(), stmt.isPasswordPlain()));
-
+                    analysePassword(stmt.getUserIdent(), stmt.getOriginalPassword(), stmt.isPasswordPlain(),
+                            stmt instanceof AlterUserStmt));
             /*
              * IDENTIFIED WITH
              */
             if (!Strings.isNullOrEmpty(stmt.getAuthPlugin())) {
+                stmt.setAuthPlugin(stmt.getAuthPlugin().toUpperCase());
                 if (AuthPlugin.AUTHENTICATION_LDAP_SIMPLE.name().equals(stmt.getAuthPlugin())) {
                     stmt.setUserForAuthPlugin(stmt.getAuthString());
                 } else if (AuthPlugin.MYSQL_NATIVE_PASSWORD.name().equals(stmt.getAuthPlugin())) {
                     // in this case, authString is password
                     stmt.setScramblePassword(analysePassword(stmt.getUserIdent(), stmt.getAuthString(),
-                            stmt.isPasswordPlain()));
-                } else if (AuthPlugin.AUTHENTICATION_KERBEROS.name().equals(stmt.getAuthPlugin()) &&
+                            stmt.isPasswordPlain(), stmt instanceof AlterUserStmt));
+                } else if (AuthPlugin.AUTHENTICATION_KERBEROS.name().equalsIgnoreCase(stmt.getAuthPlugin()) &&
                         GlobalStateMgr.getCurrentState().getAuth().isSupportKerberosAuth()) {
                     // In kerberos authentication, userForAuthPlugin represents the user principal realm.
                     // If user realm is not specified when creating user, the service principal realm will be used as
